weaviate/requirements.txt
flask==2.1.0
weaviate-client==3.3.1
openai==0.10.5
transformers==4.12.0
python-dotenv==0.19.1
